add core balance logic and use it for API balance logic
