+++ app/tfchain/api.js
@@ -3,6 +3,7 @@ import {Currency} from './tfchain.types.PrimitiveTypes.js';
 import {UnlockHash, UnlockHashType} from './tfchain.types.ConditionTypes.js';
 import * as tfwallet from './tfchain.wallet.js';
 import * as tfclient from './tfchain.client.js';
+import * as wbalance from './tfchain.balance.js';
 import * as tfexplorer from './tfchain.explorer.js';
 import * as tfnetwork from './tfchain.network.js';
 import * as tfsiabin from './tfchain.encoding.siabin.js';
@@ -488,7 +489,7 @@ export var Account =  __class__ ('Account', [object], {
 				}
 				else {
 				}
-				return Balance (self._network_type, __kwargtrans__ ({amount: 0}));
+				return Balance (self._network_type);
 			};
 			return jsasync.as_promise (no_balance_cb);
 		}
@@ -698,23 +699,23 @@ export var Wallet =  __class__ ('Wallet', [object], {
 		else {
 		}
 		var wallet = self.clone ();
-		return jsasync.as_promise (wallet.balance_sync);
-	});},
-	get _get_balance_sync () {return __get__ (this, function (self) {
-		if (arguments.length) {
-			var __ilastarg0__ = arguments.length - 1;
-			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
-				var __allkwargs0__ = arguments [__ilastarg0__--];
-				for (var __attrib0__ in __allkwargs0__) {
-					switch (__attrib0__) {
-						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+		var create_api_balance_obj = function (balance) {
+			if (arguments.length) {
+				var __ilastarg0__ = arguments.length - 1;
+				if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+					var __allkwargs0__ = arguments [__ilastarg0__--];
+					for (var __attrib0__ in __allkwargs0__) {
+						switch (__attrib0__) {
+							case 'balance': var balance = __allkwargs0__ [__attrib0__]; break;
+						}
 					}
 				}
 			}
-		}
-		else {
-		}
-		return Balance (self._network_type);
+			else {
+			}
+			return Balance (wallet._tfwallet.network_type, balance);
+		};
+		return jsasync.chain (wallet._tfwallet.balance, create_api_balance_obj);
 	});},
 	get transaction_new () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -733,7 +734,6 @@ export var Wallet =  __class__ ('Wallet', [object], {
 		return CoinTransactionBuilder (self);
 	});}
 });
-Object.defineProperty (Wallet, 'balance_sync', property.call (Wallet, Wallet._get_balance_sync));
 Object.defineProperty (Wallet, 'balance', property.call (Wallet, Wallet._get_balance));
 Object.defineProperty (Wallet, 'address_count', property.call (Wallet, Wallet._get_address_count));
 Object.defineProperty (Wallet, 'addresses', property.call (Wallet, Wallet._get_addresses));
@@ -835,30 +835,32 @@ export var CoinTransactionBuilder =  __class__ ('CoinTransactionBuilder', [objec
 		else {
 		}
 		var wallet = self._wallet.clone ();
-		var cb = function () {
+		var cb = function (balance) {
 			if (arguments.length) {
 				var __ilastarg0__ = arguments.length - 1;
 				if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
 					var __allkwargs0__ = arguments [__ilastarg0__--];
 					for (var __attrib0__ in __allkwargs0__) {
+						switch (__attrib0__) {
+							case 'balance': var balance = __allkwargs0__ [__attrib0__]; break;
+						}
 					}
 				}
 			}
 			else {
 			}
-			var balance = wallet.balance_sync;
 			print ('Sent from wallet {} succesfully with an input balance of {} TFT!'.format (wallet.wallet_name, balance.coins_total));
 			return '66ccdf3a0bca58025be7fdc71f3f6bfbd6ed6287aa698a131734a947c71a3bbf';
 		};
 		print ('Sending from wallet {}...'.format (wallet.wallet_name));
-		return jsasync.chain (jsasync.sleep (3000), cb);
+		return jsasync.chain (wallet.balance, cb);
 	});}
 });
 export var Balance =  __class__ ('Balance', [object], {
 	__module__: __name__,
-	get __init__ () {return __get__ (this, function (self, network_type, amount) {
-		if (typeof amount == 'undefined' || (amount != null && amount.hasOwnProperty ("__kwargtrans__"))) {;
-			var amount = null;
+	get __init__ () {return __get__ (this, function (self, network_type, tfbalance) {
+		if (typeof tfbalance == 'undefined' || (tfbalance != null && tfbalance.hasOwnProperty ("__kwargtrans__"))) {;
+			var tfbalance = null;
 		};
 		if (arguments.length) {
 			var __ilastarg0__ = arguments.length - 1;
@@ -868,25 +870,30 @@ export var Balance =  __class__ ('Balance', [object], {
 					switch (__attrib0__) {
 						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
 						case 'network_type': var network_type = __allkwargs0__ [__attrib0__]; break;
-						case 'amount': var amount = __allkwargs0__ [__attrib0__]; break;
+						case 'tfbalance': var tfbalance = __allkwargs0__ [__attrib0__]; break;
 					}
 				}
 			}
 		}
 		else {
 		}
-		if (amount === null) {
-			self._amount = 1;
+		if (!(isinstance (network_type, tfnetwork.Type))) {
+			var __except0__ = py_TypeError ('network_type has to be of type tfchain.network.Type, not be of type {}'.format (py_typeof (network_type)));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		self._network_type = network_type;
+		if (tfbalance === null) {
+			self._tfbalance = wbalance.WalletBalance ();
 		}
 		else {
-			if (!(isinstance (amount, int))) {
-				var __except0__ = py_TypeError ('amount can only be int or None, not be of type {}'.format (py_typeof (amount)));
+			if (!(isinstance (tfbalance, wbalance.WalletBalance))) {
+				var __except0__ = py_TypeError ('tfbalance has to be of type tfchain.balance.WalletBalance, not be of type {}'.format (py_typeof (tfbalance)));
 				__except0__.__cause__ = null;
 				throw __except0__;
 			}
-			self._amount = max (amount, 0);
+			self._tfbalance = tfbalance;
 		}
-		self._network_type = network_type;
 	});},
 	get merge () {return __get__ (this, function (self, other) {
 		if (arguments.length) {
@@ -908,7 +915,7 @@ export var Balance =  __class__ ('Balance', [object], {
 			__except0__.__cause__ = null;
 			throw __except0__;
 		}
-		return Balance (__kwargtrans__ ({network_type: self._network_type, amount: self._amount + other._amount}));
+		return Balance (__kwargtrans__ ({network_type: self._network_type, tfbalance: wbalance.WalletBalance ().balance_add (self._tfbalance).balance_add (other._tfbalance)}));
 	});},
 	get spend_amount_is_valid () {return __get__ (this, function (self, amount) {
 		if (arguments.length) {
@@ -957,7 +964,7 @@ export var Balance =  __class__ ('Balance', [object], {
 		}
 		else {
 		}
-		return Currency (jsstr.from_int (self._amount));
+		return self._tfbalance.available;
 	});},
 	get _get_coins_locked () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -973,7 +980,7 @@ export var Balance =  __class__ ('Balance', [object], {
 		}
 		else {
 		}
-		return Currency (jsstr.from_int (self._amount));
+		return self._tfbalance.locked;
 	});},
 	get _get_coins_total () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -1007,7 +1014,7 @@ export var Balance =  __class__ ('Balance', [object], {
 		else {
 		}
 		UnlockHash.from_str (address);
-		return Balance (self._network_type, __mod__ (jshex.hex_to_int (address [3]), 9) + 1);
+		return Balance (self._network_type, self._tfbalance);
 	});},
 	get _get_transactions () {return __get__ (this, function (self) {
 		if (arguments.length) {
