+++ app/tfchain/tfchain.errors.js
@@ -1,5 +1,6 @@
 import {AssertionError, AttributeError, BaseException, DeprecationWarning, Exception, IndexError, IterableError, KeyError, NotImplementedError, RuntimeWarning, StopIteration, UserWarning, ValueError, Warning, __JsIterator__, __PyIterator__, __Terminal__, __add__, __and__, __call__, __class__, __envir__, __eq__, __floordiv__, __ge__, __get__, __getcm__, __getitem__, __getslice__, __getsm__, __gt__, __i__, __iadd__, __iand__, __idiv__, __ijsmod__, __ilshift__, __imatmul__, __imod__, __imul__, __in__, __init__, __ior__, __ipow__, __irshift__, __isub__, __ixor__, __jsUsePyNext__, __jsmod__, __k__, __kwargtrans__, __le__, __lshift__, __lt__, __matmul__, __mergefields__, __mergekwargtrans__, __mod__, __mul__, __ne__, __neg__, __nest__, __or__, __pow__, __pragma__, __proxy__, __pyUseJsNext__, __rshift__, __setitem__, __setproperty__, __setslice__, __sort__, __specialattrib__, __sub__, __super__, __t__, __terminal__, __truediv__, __withblock__, __xor__, abs, all, any, assert, bool, bytearray, bytes, callable, chr, copy, deepcopy, delattr, dict, dir, divmod, enumerate, filter, float, getattr, hasattr, input, int, isinstance, issubclass, len, list, map, max, min, object, ord, pow, print, property, py_TypeError, py_iter, py_metatype, py_next, py_reversed, py_typeof, range, repr, round, set, setattr, sorted, str, sum, tuple, zip} from './org.transcrypt.__runtime__.js';
 import {datetime} from './datetime.js';
+import * as jsobj from './tfchain.polyfill.encoding.object.js';
 var __name__ = 'tfchain.errors';
 export var InvalidPublicKeySpecifier =  __class__ ('InvalidPublicKeySpecifier', [Exception], {
 	__module__: __name__,
@@ -151,7 +152,13 @@ export var ExplorerError =  __class__ ('ExplorerError', [Exception], {
 	});}
 });
 Object.defineProperty (ExplorerError, 'endpoint', property.call (ExplorerError, ExplorerError._get_endpoint));;
-export var ExplorerNoContent =  __class__ ('ExplorerNoContent', [ExplorerError], {
+export var ExplorerUserError =  __class__ ('ExplorerUserError', [ExplorerError], {
+	__module__: __name__,
+});
+export var ExplorerNoContent =  __class__ ('ExplorerNoContent', [ExplorerUserError], {
+	__module__: __name__,
+});
+export var ExplorerBadRequest =  __class__ ('ExplorerBadRequest', [ExplorerUserError], {
 	__module__: __name__,
 });
 export var ExplorerServerError =  __class__ ('ExplorerServerError', [ExplorerError], {
@@ -262,7 +269,7 @@ export var ExplorerInvalidResponse =  __class__ ('ExplorerInvalidResponse', [Exp
 		else {
 		}
 		__super__ (ExplorerInvalidResponse, '__init__') (self, message, endpoint);
-		if (!(isinstance (response, dict))) {
+		if (!(isinstance (response, dict)) && !(jsobj.is_js_obj (response))) {
 			var __except0__ = py_TypeError ('invalid response, expected it to be of type dict not {}'.format (py_typeof (response)));
 			__except0__.__cause__ = null;
 			throw __except0__;
