+++ app/tfchain/tfchain.wallet.py
@@ -6,6 +6,7 @@ import tfchain.client as tfclient
 import tfchain.errors as tferrors
 
 from tfchain.network import Type as NetworkType
+from tfchain.balance import WalletsBalance
 
 from tfchain.types import ConditionTypes, transactions
 from tfchain.types.IO import CoinInput
@@ -13,8 +14,6 @@ from tfchain.types.CryptoTypes import PublicKey, PublicKeySpecifier
 from tfchain.types.PrimitiveTypes import Hash, Currency
 from tfchain.types.ConditionTypes import UnlockHash, UnlockHashType, ConditionMultiSignature
 from tfchain.types.FulfillmentTypes import FulfillmentMultiSignature, PublicKeySignaturePair
-
-_MAX_RIVINE_TRANSACTION_INPUTS = 99
 
 class TFChainWallet:
     """
@@ -141,53 +140,14 @@ class TFChainWallet:
         balance = self.balance
         return balance.addresses_multisig
 
-    # TODO: fix this code: currently this is not working,
-    # need to port it to the promise-based logic
-
     @property
     def balance(self):
         """
         The balance "sheet" of the wallet.
         """
-        # first get chain info
-        info = self._client.blockchain_info_get()
-
-        addresses = self.addresses
-        balance = WalletsBalance()
-        # collect info for all personal addresses
-        multisig_addresses = []
-        for address in addresses:
-            try:
-                # collect the inputs/outputs linked to this address for all found transactions
-                result = self._unlockhash_get(address)
-                uh_balance = result.balance(info=info)
-                balance = balance.balance_add(uh_balance)
-                # collect all multisig addresses
-                for address in result.multisig_addresses:
-                    multisig_addresses.append(address.__str__())
-            except tferrors.ExplorerNoContent:
-                 # ignore this exception as it simply means
-                 # the address has no activity yet on the chain
-                pass
-
-        # collect info for all multisig addresses
-        for address in multisig_addresses:
-            try:
-                # collect the inputs/outputs linked to this address for all found transactions
-                result = self._unlockhash_get(address)
-                uh_balance = result.balance(info=info)
-                balance = balance.balance_add(uh_balance)
-            except tferrors.ExplorerNoContent:
-                 # ignore this exception as it simply means
-                 # the address has no activity yet on the chain
-                pass
-        # ensure info is defined for wallet, even if no content
-        balance.chain_blockid = info.blockid
-        balance.chain_time = info.timestamp
-        balance.chain_height = info.height
-        # return the balance
-        return balance
-
+        w = self.clone()
+        aggregator = WalletBalanceAggregator(w)
+        return aggregator.fetch_and_aggregate()
 
     @property
     def transactions(self):
@@ -200,7 +160,10 @@ class TFChainWallet:
             for address in w.addresses:
                 yield w._unlockhash_get(address)
         transactions = set()
-        p = jsasync.promise_pool_new(generator, cb=(lambda result: transactions.update(result.transactions)))
+        def gatherer(result):
+            if result.transactions:
+                transactions.update(result.transactions)
+        p = jsasync.promise_pool_new(generator, cb=gatherer)
 
         # define sort cb that will sort it prior to the final return
         def cb():
@@ -1622,597 +1585,65 @@ class TransactionSendResult():
 #     def submitted(self):
 #         return self._submitted
 
-class WalletBalance:
-    def __init__(self):
-        # personal wallet outputs
-        self._outputs = {}
-        self._outputs_spent = {}
-        self._outputs_unconfirmed = {}
-        self._outputs_unconfirmed_spent = {}
-        # balance chain context
-        self._chain_time = 0
-        self._chain_height = 0
-        self._chain_blockid = Hash()
-        # all wallet addresses tracked in this wallet
-        self._addresses = set()
+class WalletBalanceAggregator:
+    """
+    State class to serve as the red line throughout
+    the chained promise-based balance gathering for a wallet,
+    which can involve the merging of results of multiple addresses,
+    single- and/or multisignature.
+    """
 
-    @property
-    def addresses(self):
-        """
-        All (personal wallet) addresses for which an output is tracked in this Balance.
-        """
-        return list(self._addresses)
+    def __init__(self, wallet):
+        self._wallet = wallet
+        self._balance = WalletsBalance()
+        self._multisig_addresses = []
+        self._info = None
 
-    @property
-    def chain_blockid(self):
-        """
-        Blockchain block ID, as defined by the last known block.
-        """
-        return self._chain_blockid
-    @chain_blockid.setter
-    def chain_blockid(self, value):
-        """
-        Set the blockchain block ID, such that applications that which to cache this
-        balance object could ensure that the last block is still the same as the
-        last known block known by this balance instance.
-        """
-        if not value:
-            self._chain_blockid = Hash()
-            return
-        if isinstance(value, Hash):
-            self._chain_blockid.value = value.value
-        else:
-            self._chain_blockid.value = value
+    def fetch_and_aggregate(self):
+        return jsasync.chain(
+            self._wallet._client.blockchain_info_get(),
+            self._collect_chain_info,
+            self._personal_pool_chain_get,
+            self._multisig_pool_chain_get,
+            self._balance_get,
+        )
 
-    @property
-    def chain_time(self):
-        """
-        Blockchain time, as defined by the last known block.
-        """
-        return self._chain_time
-    @chain_time.setter
-    def chain_time(self, value):
-        """
-        Set the blockchain time, such that the balance object can report
-        locked/unlocked outputs correctly for outputs that are locked by time.
-        """
-        if not isinstance(value, int):
-            raise TypeError("WalletBalance's chain time cannot be of type {} (expected: int)".format(type(value)))
-        self._chain_time = int(value)
+    def _collect_chain_info(self, info):
+        self._info = info
 
-    @property
-    def chain_height(self):
-        """
-        Blockchain height, as defined by the last known block.
-        """
-        return self._chain_height
-    @chain_height.setter
-    def chain_height(self, value):
-        """
-        Set the blockchain height, such that the balance object can report
-        locked/unlocked outputs correctly for outputs that are locked by height.
-        """
-        if not isinstance(value, int):
-            raise TypeError("WalletBalance's chain height cannot be of type {} (expected: int)".format(type(value)))
-        self._chain_height = int(value)
+    def _personal_pool_chain_get(self):
+        return jsasync.promise_pool_new(
+            self._personal_address_generator,
+            cb=self._collect_personal_balance,
+        )
 
-    @property
-    def active(self):
-        """
-        Returns if this balance is active,
-        meaning it has available outputs to spend (confirmed or not).
-        """
-        return len(self._outputs) > 0 or len(self._outputs_unconfirmed) > 0
+    def _personal_address_generator(self):
+        for address in self._wallet.addresses:
+            yield self._wallet._unlockhash_get(address)
 
-    @property
-    def outputs_spent(self):
-        """
-        Spent (coin) outputs.
-        """
-        return self._outputs_spent
-    @property
-    def outputs_unconfirmed(self):
-        """
-        Unconfirmed (coin) outputs, available for spending or locked.
-        """
-        return self._outputs_unconfirmed
-    @property
-    def outputs_unconfirmed_available(self):
-        """
-        Unconfirmed (coin) outputs, available for spending.
-        """
-        if self.chain_time > 0 and self.chain_height > 0:
-            return [co for co in self._outputs_unconfirmed.values()
-                if not co.condition.lock.locked_check(time=self.chain_time, height=self.chain_height)]
-        else:
-            return list(self._outputs_unconfirmed.values())
-    @property
-    def outputs_unconfirmed_spent(self):
-        """
-        Unconfirmed (coin) outputs that have already been spent.
-        """
-        return self._outputs_unconfirmed_spent
+    def _collect_personal_balance(self, result):
+        balance = result.balance(info=self._info)
+        self._balance = self._balance.balance_add(balance)
+        # collect all multisig addresses
+        for address in result.multisig_addresses:
+            self._multisig_addresses.append(address.__str__())
 
-    @property
-    def outputs_available(self):
-        """
-        Total available (coin) outputs.
-        """
-        if self.chain_time > 0 and self.chain_height > 0:
-            return [co for co in self._outputs.values()
-                if not co.condition.lock.locked_check(time=self.chain_time, height=self.chain_height)]
-        else:
-            return jsobj.dict_values(self._outputs)
+    def _multisig_pool_chain_get(self):
+        return jsasync.promise_pool_new(
+            self._multisig_address_generator,
+            cb=self._collect_multisig_balance,
+        )
 
-    @property
-    def available(self):
-        """
-        Total available coins.
-        """
-        return Currency.sum(*[co.value for co in self.outputs_available])
+    def _multisig_address_generator(self):
+        for address in self._multisig_addresses:
+            yield self._wallet._unlockhash_get(address)
 
-    @property
-    def locked(self):
-        """
-        Total available coins that are locked.
-        """
-        if self.chain_time > 0 and self.chain_height > 0:
-            return Currency.sum(*[co.value for co in jsobj.dict_values(self._outputs)
-                if co.condition.lock.locked_check(time=self.chain_time, height=self.chain_height)]) or Currency()
-        else:
-            return Currency() # impossible to know for sure without a complete context
+    def _collect_multisig_balance(self, result):
+        balance = result.balance(info=self._info)
+        self._balance = self._balance.balance_add(balance)
 
-    @property
-    def unconfirmed(self):
-        """
-        Total unconfirmed coins, available for spending.
-        """
-        if self.chain_time > 0 and self.chain_height > 0:
-            return Currency.sum(*[co.value for co in jsobj.dict_values(self._outputs_unconfirmed)
-                if not co.condition.lock.locked_check(time=self.chain_time, height=self.chain_height)]) or Currency()
-        else:
-            return Currency.sum(*[co.value for co in jsobj.dict_values(self._outputs_unconfirmed)])
-
-    @property
-    def unconfirmed_locked(self):
-        """
-        Total unconfirmed coins that are locked, and thus not available for spending.
-        """
-        if self.chain_time > 0 and self.chain_height > 0:
-            return Currency.sum(*[co.value for co in jsobj.dict_values(self._outputs_unconfirmed)
-                if co.condition.lock.locked_check(time=self.chain_time, height=self.chain_height)]) or Currency()
-        else:
-            return Currency() # impossible to know for sure without a complete context
-
-    def output_add(self, output, confirmed=True, spent=False):
-        """
-        Add an output to the Wallet's balance.
-        """
-        strid = output.id.__str__()
-        if confirmed: # confirmed outputs
-            if spent:
-                self._outputs_spent[strid] = output
-                # delete from other output lists if prior registered
-                self._outputs.pop(strid, None)
-                self._outputs_unconfirmed.pop(strid, None)
-                self._outputs_unconfirmed_spent.pop(strid, None)
-            elif strid not in self._outputs_spent and strid not in self._outputs_unconfirmed_spent:
-                self._outputs[strid] = output
-                # delete from other output lists if prior registered
-                self._outputs_unconfirmed.pop(strid, None)
-        elif strid not in self._outputs_spent: # unconfirmed outputs
-            if spent:
-                self._outputs_unconfirmed_spent[strid] = output
-                # delete from other output lists if prior registered
-                self._outputs_unconfirmed.pop(strid, None)
-                self._outputs.pop(strid, None)
-            elif strid not in self._outputs_unconfirmed_spent:
-                self._outputs_unconfirmed[strid] = output
-        self._addresses.add(str(output.condition.unlockhash))
-
-    @property
-    def _base(self):
-        """
-        Private helper utility to return this class as a new and pure WalletBalance object
-        """
-        b = WalletBalance()
-        b._outputs = self._outputs
-        b._outputs_spent = self._outputs_spent
-        b._outputs_unconfirmed = self._outputs_unconfirmed
-        b._outputs_unconfirmed_spent = self._outputs_unconfirmed_spent
-        b._chain_blockid = self._chain_blockid
-        b._chain_height = self._chain_height
-        b._chain_time = self._chain_time
-        b._addresses = self._addresses
-        return b
-
-    def balance_add(self, other):
-        """
-        Merge the content of the other balance into this balance.
-        If other is None, this call results in a no-op.
-
-        Always assign the result, as it could other than self,
-        should the class type be changed in order to add all content correctly.
-        """
-        if other is None:
-            return self
-        if isinstance(other, (WalletsBalance, MultiSigWalletBalance)):
-            return WalletsBalance().balance_add(self).balance_add(other)
-        if not isinstance(other, WalletBalance):
-            raise TypeError("other balance has to be of type wallet balance")
-        # another balance is defined, create a new balance that will contain our merge
-        # merge the chain info
-        if self.chain_height >= other.chain_height:
-            if self.chain_time < other.chain_time:
-                raise ValueError("chain time and chain height of balances do not match")
-        else:
-            if self.chain_time >= other.chain_time:
-                raise ValueError("chain time and chain height of balances do not match")
-            self.chain_time = other.chain_time
-            self.chain_height = other.chain_height
-            self.chain_blockid = other.chain_blockid
-        # merge the outputs
-        for attr in ['_outputs', '_outputs_spent', '_outputs_unconfirmed', '_outputs_unconfirmed_spent']:
-            d = jsobj.get_attr(self, attr, {})
-            for (id, output) in jsobj.get_items(jsobj.get_attr(other, attr, {})):
-                d[id] = output
-        # merge the addresses
-        self._addresses |= other._addresses
-        # return the modified self
-        return self
-
-    def drain(self, recipient, miner_fee, unconfirmed=False, data=None, lock=None):
-        """
-        add all available outputs into as many transactions as required,
-        by default only confirmed outputs are used, if unconfirmed=True
-        it will use unconfirmed available outputs as well.
-
-        Result can be an empty list if no outputs were available.
-
-        @param recipient: required recipient towards who the drained coins will be sent
-        @param the miner fee to be added to all sent transactions
-        @param unconfirmed: optionally drain unconfirmed (available) outputs as well
-        @param data: optional data that can be attached ot the created transactions (str or bytes), with a max length of 83
-        @param lock: optional lock that can be attached to the sent coin outputs
-        """
-        # define recipient
-        recipient = ConditionTypes.from_recipient(recipient, lock=lock)
-
-        # validate miner fee
-        if not isinstance(miner_fee, Currency):
-            raise TypeError("miner fee has to be a currency")
-        if miner_fee.__eq__(0):
-            raise ValueError("a non-zero miner fee has to be defined")
-
-        # collect all transactions in one list
-        txns = []
-
-        # collect all confirmed (available) outputs
-        outputs = self.outputs_available
-        if unconfirmed:
-            # if also the unconfirmed_avaialble) outputs are desired, let's add them as well
-            outputs += self.outputs_unconfirmed_available
-        # drain all outputs
-        while len(outputs) > 0:
-            txn = transactions.new()
-            txn.data = data
-            txn.miner_fee_add(miner_fee)
-            # select maximum _MAX_RIVINE_TRANSACTION_INPUTS outputs
-            n = min(len(outputs), _MAX_RIVINE_TRANSACTION_INPUTS)
-            used_outputs = outputs[:n]
-            outputs = outputs[n:] # and update our output list, so we do not double spend
-            # compute amount, minus minimum fee and add our only output
-            amount = sum([co.value for co in used_outputs]) - miner_fee
-            txn.coin_output_add(condition=recipient, value=amount)
-            # add the coin inputs
-            txn.coin_inputs = [CoinInput.from_coin_output(co) for co in used_outputs]
-            # append the transaction
-            txns.append(txn)
-
-        # return all created transactions, if any
-        return txns
-
-class MultiSigWalletBalance(WalletBalance):
-    def __init__(self, owners, signature_count):
-        """
-        Creates a personal multi signature wallet.
-        """
-        if not isinstance(signature_count, int):
-            raise TypeError("signature count of a MultiSigWallet is expected to be of type int, not {}".format(type(signature_count)))
-        if signature_count < 1:
-            raise ValueError("signature count of a MultiSigWallet has to be at least 1, cannot be {}".format(signature_count))
-        if len(owners) < signature_count:
-            raise ValueError("the amount of owners ({}) cannot be lower than the specified signature count ({})".format(len(owners), signature_count))
-        self._owners = owners
-        self._signature_count = signature_count
-        super().__init__()
-
-    @property
-    def address(self):
-        """
-        The address of this MultiSig Wallet
-        """
-        return ConditionMultiSignature(unlockhashes=self._owners, min_nr_sig=self._signature_count).unlockhash.__str__()
-
-    @property
-    def owners(self):
-        """
-        Owners of this MultiSignature Wallet
-        """
-        return [owner.__str__() for owner in self._owners]
-
-    @property
-    def signature_count(self):
-        """
-        Amount of signatures minimum required
-        """
-        return self._signature_count
-
-    def balance_add(self, other):
-        """
-        Merge the content of the other balance into this balance.
-        If other is None, this call results in a no-op.
-
-        Always assign the result, as it could other than self,
-        should the class type be changed in order to add all content correctly.
-        """
-        if other is None:
-            return self
-        if not isinstance(other, MultiSigWalletBalance):
-            if isinstance(other, (WalletBalance, WalletsBalance)):
-                return WalletsBalance().balance_add(self).balance_add(self)
-            # can only merge 2 multi-signature wallet balances
-            raise TypeError("other balance has to be of type multi-signature wallet balance")
-        if self.address != other.addres:
-            raise ValueError("other balance is for a different MultiSignature Wallet, cannot be merged")
-        # piggy-back on the super class for the actual merge logic
-        return super().balance_add(other._base)
-
-class WalletsBalance(WalletBalance):
-    def __init__(self):
-        """
-        Creates a personal wallet, which also can have children wallets that are meant for
-        all MultiSignature wallets that are related to one or more addresses of the personal wallet.
-        """
-        self._wallets = {}
-        super().__init__()
-
-    @property
-    def wallets(self):
-        """
-        All multisig wallets linked to this wallet.
-        """
-        return self._wallets
-
-    @property
-    def addresses_multisig(self):
-        """
-        All (multisig wallet) addresses for which an output is tracked in this Balance.
-        For each address you'll find a wallet in the `self.wallets` dict property.
-        """
-        return jsobj.get_keys(self.wallets)
-
-    def multisig_output_add(self, address, output, confirmed=True, spent=False):
-        """
-        Add an output to the MultiSignature Wallet's balance.
-        """
-        oc = output.condition.unwrap()
-        if not isinstance(oc, ConditionMultiSignature):
-            raise TypeError("multi signature's output condition cannot be of type {} (expected: ConditionMultiSignature)".format(type(oc)))
-        if not address in self._wallets:
-            self._wallets[address] = MultiSigWalletBalance(
-                owners=output.condition.unlockhashes,
-                signature_count=output.condition.required_signatures)
-        self._wallets[address].output_add(output, confirmed=confirmed, spent=spent)
-
-    def output_add(self, output, confirmed=True, spent=False):
-        """
-        Add an output to the Wallet's balance.
-        """
-        uh = output.condition.unlockhash
-        if uh.type.__eq__(UnlockHashType.MULTI_SIG):
-            return self.multisig_output_add(address=uh.__str__(), output=output, confirmed=confirmed, spent=spent)
-        self._addresses.add(uh.__str__())
-        return super().output_add(output=output, confirmed=confirmed, spent=spent)
-
-    def balance_add(self, other):
-        """
-        Merge the content of the other balance into this balance.
-        If other is None, this call results in a no-op.
-
-        Always assign the result, as it could other than self,
-        should the class type be changed in order to add all content correctly.
-        """
-        if other is None:
-            return self
-        if not isinstance(other, WalletBalance):
-            raise TypeError("other balance has to be of type wallet balance")
-        if isinstance(other, MultiSigWalletBalance):
-            self._merge_multisig_wallet_balance(other.address, other)
-            return self
-        # piggy-back on the super class for the actual output merge logic
-        b = super().balance_add(other._base)
-        if b != self:
-            raise Exception("BUG: instance shouldn't have changed, please fix or report")
-        if not isinstance(other, WalletsBalance):
-            return b # return as nothing else can be merged
-        # merge all the multi-signature wallets
-        for (addr, balance) in jsobj.get_items(other._wallets):
-            b._merge_multisig_wallet_balance(addr, balance)
-        # return the merged balance
-        return b
-
-    def _merge_multisig_wallet_balance(self, address, balance):
-        """
-        Assign or merge a multi-sig wallet balance
-        """
-        if address not in self._wallets:
-            self._wallets[address] = balance
-            return
-        self._wallets[address] = self._wallets[address].merge(balance)
-
-    def fund(self, amount, source=None):
-        """
-        Fund the specified amount with the available outputs of this wallet's balance.
-        """
-        # collect addresses and multisig addresses
-        addresses = set()
-        ms_addresses = set()
-        refund = None
-        if source is None:
-            for co in self.outputs_available:
-                addresses.add(co.condition.unlockhash.__str__())
-            for co in self.outputs_unconfirmed_available:
-                addresses.add(co.condition.unlockhash.__str__())
-        else:
-            # if only one address is given, transform it into an acceptable list
-            if not isinstance(source, list) and not jsobj.is_js_arr(source):
-                if isinstance(source, str):
-                    source = UnlockHash.from_json(source)
-                elif not isinstance(source, UnlockHash):
-                    raise TypeError("cannot add source address from type {}".format(type(source)))
-                source = [source]
-            # add one or multiple personal/multisig addresses
-            for value in source:
-                if isinstance(value, str):
-                    value = UnlockHash.from_json(value)
-                elif not isinstance(value, UnlockHash):
-                    raise TypeError("cannot add source address from type {}".format(type(value)))
-                if value.type.__eq__(UnlockHashType.MULTI_SIG):
-                    ms_addresses.add(value)
-                elif value.type.__eq__(UnlockHashType.PUBLIC_KEY):
-                    addresses.add(value)
-                else:
-                    raise TypeError("cannot add source addres with unsupported UnlockHashType {}".format(value.type))
-            if len(source) == 1:
-                if source[0].type.__eq__(UnlockHashType.PUBLIC_KEY):
-                    refund = ConditionTypes.unlockhash_new(unlockhash=source[0])
-                else:
-                    addr = source[0].__str__()
-                    if addr in self.wallets:
-                        wallet = self.wallets[addr]
-                        refund = ConditionTypes.multi_signature_new(min_nr_sig=wallet.signature_count, unlockhashes=wallet.owners)
-
-
-        # ensure at least one address is defined
-        if len(addresses) == 0 and len(ms_addresses) == 0:
-            raise tferrors.InsufficientFunds("insufficient funds in this wallet")
-
-        # if personal addresses are given, try to use these first
-        # as these are the easiest kind to deal with
-        if len(addresses) == 0:
-            outputs, collected = ([], Currency()) # start with nothing
-        else:
-            outputs, collected = self._fund_individual(amount, addresses)
-
-        if collected.greater_than_or_equal_to(amount):
-            # if we already have sufficient, we stop now
-            return ([CoinInput.from_coin_output(co) for co in outputs], collected.minus(amount), refund)
-
-        if len(ms_addresses) == 0:
-            # if no ms_addresses were defined,
-            raise tferrors.InsufficientFunds("not enough funds available in the individual wallet to fund the requested amount")
-        # otherwise keep going for multisig addresses
-        outputs, collected = self._fund_multisig(amount, ms_addresses, outputs=outputs, collected=collected)
-
-        # if we still didn't manage to fund enough, there is nothing we can do
-        if collected.less_than(amount):
-            raise tferrors.InsufficientFunds("not enough funds available in the wallets to fund the requested amount")
-        return ([CoinInput.from_coin_output(co) for co in outputs], collected.minus(amount), refund)
-
-    def _fund_individual(self, amount, addresses):
-        outputs_available = [co for co in self.outputs_available if co.condition.unlockhash in addresses]
-        def sort_output_by_value(a, b):
-            if a.value.less_than(b.value):
-                return -1
-            if a.value.greater_than(b.value):
-                return 1
-            return 0
-        outputs_available = jsarr.sort(outputs_available, sort_output_by_value)
-        collected = Currency()
-        outputs = []
-        # try to fund only with confirmed outputs, if possible
-        for co in outputs_available:
-            if co.value.greater_than_or_equal_to(amount):
-                return [co], co.value
-            collected = collected.plus(co.value)
-            outputs.append(co)
-            if len(outputs) > _MAX_RIVINE_TRANSACTION_INPUTS:
-                # to not reach the input limit
-                collected = collected.minus(jsarr.pop(outputs, 0).value)
-            if collected.greater_than_or_equal_to(amount):
-                return outputs, collected
-
-        if collected.greater_than_or_equal_to(amount):
-            # if we already have sufficient, we stop now
-            return outputs, collected
-
-        # use unconfirmed balance, not ideal, but acceptable
-        outputs_available = [co for co in self.outputs_unconfirmed_available if co.condition.unlockhash in addresses]
-        outputs_available = jsarr.sort(outputs_available, sort_output_by_value, reverse=True)
-        for co in outputs_available:
-            if co.value.greater_than_or_equal_to(amount):
-                return [co], co.value
-            collected = collected.plus(co.value)
-            outputs.append(co)
-            if len(outputs) > _MAX_RIVINE_TRANSACTION_INPUTS:
-                # to not reach the input limit
-                collected = collected.minus(outputs.pop(0).value)
-            if collected.greater_than_or_equal_to(amount):
-                return outputs, collected
-
-        # we return whatever we have collected, no matter if it is sufficient
-        return outputs, collected
-
-    def _fund_multisig(self, amount, addresses, outputs=None, collected=None):
-        if outputs is None:
-            outputs = []
-        if collected is None:
-            collected = Currency()
-        for address, wallet in self.wallets.items():
-            if UnlockHash.from_json(address).__str__() not in addresses:
-                continue # nothing to do here
-
-            outputs_available = wallet.outputs_available
-            def sort_output_by_value(a, b):
-                if a.value.less_than(b.value):
-                    return -1
-                if a.value.greater_than(b.value):
-                    return 1
-                return 0
-            outputs_available = jsarr.sort(outputs_available, sort_output_by_value)
-            # try to fund only with confirmed outputs, if possible
-            for co in outputs_available:
-                if co.value.greater_than_or_equal_to(amount):
-                    return [co], co.value
-
-                collected = collected.plus(co.value)
-                outputs.append(co)
-                if len(outputs) > _MAX_RIVINE_TRANSACTION_INPUTS:
-                    # to not reach the input limit
-                    collected = collected.minus(outputs.pop(0).value)
-                if collected.greater_than_or_equal_to(amount):
-                    return outputs, collected
-
-            if collected.greater_than_or_equal_to(amount):
-                # if we already have sufficient, we stop now
-                return outputs, collected
-
-            # use unconfirmed balance, not ideal, but acceptable
-            outputs_available = wallet.outputs_unconfirmed_available
-            outputs_available = jsarr.sort(outputs_available, sort_output_by_value, reverse=True)
-            for co in outputs_available:
-                if co.value.greater_than_or_equal_to(amount):
-                    return [co], co.value
-                collected = collected.plus(co.value)
-                outputs.append(co)
-                if len(outputs) > _MAX_RIVINE_TRANSACTION_INPUTS:
-                    # to not reach the input limit
-                    collected = collected.minus(outputs.pop(0).value)
-                if collected.greater_than_or_equal_to(amount):
-                    return outputs, collected
-        # we return whatever we have collected, no matter if it is sufficient
-        return outputs, collected
+    def _balance_get(self):
+        return self._balance
 
 class CoinTransactionBuilder():
     def __init__(self, wallet):
