+++ src/tfchain/polyfill/http.py
@@ -19,10 +19,14 @@ def http_get(resource, headers=None):
                 };
             });
         }
-        return response.json().then(function(data) {
+        return response.text().then(function(data) {
+            let message;
+            try {
+                message = JSON.parse(data).message;
+            } catch(e) {}
             return {
                 code: response.status,
-                data: data.message || ("GET request to " + resource + " failed with status code " + response.status),
+                data: message || ("GET request to " + resource + " failed with status code " + response.status),
             };
         });
     });
